requirements.txt
asgiref==3.8.1
certifi==2024.2.2
charset-normalizer==3.3.2
crispy-bootstrap4==2024.1
Django==4.2.5
django-crispy-forms==2.1
django-debug-toolbar==4.3.0
django-tinymce==3.6.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
Faker==21.0.0
idna==3.6
Pillow==9.5.0
PyJWT==2.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
transliterate==1.10.2
typing_extensions==4.10.0
urllib3==1.26.18
